// src/app/api/generated/model/boardMemberResponseDto.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Boards API
 * REST API for Boards, Spring Boot version
 * OpenAPI spec version: 1.0.0
 */
import type { BoardRoleSummaryResponseDto } from './boardRoleSummaryResponseDto';
import type { UserSummaryResponseDto } from './userSummaryResponseDto';

export interface BoardMemberResponseDto {
  user?: UserSummaryResponseDto;
  boardRole?: BoardRoleSummaryResponseDto;
  owner?: boolean;
}